requirements.txt
affine==2.2.2
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
config==0.3.9
cycler==0.10.0
dj-database-url==0.5.0
Django==2.0
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
numpy==1.16.4
Pillow==6.0.0
psycopg2==2.8.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
rasterio==1.0.24
requests==2.22.0
retrying==1.3.3
six==1.12.0
snuggs==1.4.6
urllib3==1.25.3
whitenoise==4.1.2
